Updated prepare script

+++ prepare.sh
@@ -9,18 +9,27 @@ sudo apt-get -y install \
   apt-transport-https \
   ca-certificates
 
-# Blacklist nouveau and rebuild kernel initramfs
-echo "blacklist nouveau
-options nouveau modeset=0" >> blacklist-nouveau.conf
-sudo mv blacklist-nouveau.conf /etc/modprobe.d/blacklist-nouveau.conf
-sudo update-initramfs -u
-sudo reboot
+if ! [ -f /etc/modprobe.d/nouveau-blacklist.conf ]; then
+  echo "nouveau is not blacklisted, doing so and rebooting"
 
-# Install NVIDIA Linux toolkit 510.54
-wget https://us.download.nvidia.com/XFree86/Linux-x86_64/510.54/NVIDIA-Linux-x86_64-510.54.run
-chmod +x NVIDIA-Linux-x86_64-510.54.run
-sudo bash ./NVIDIA-Linux-x86_64-510.54.run
-rm NVIDIA-Linux-x86_64-510.54.run
+  # Blacklist nouveau and rebuild kernel initramfs
+  echo "blacklist nouveau
+options nouveau modeset=0" >> blacklist-nouveau.conf
+  sudo mv blacklist-nouveau.conf /etc/modprobe.d/blacklist-nouveau.conf
+  sudo update-initramfs -u
+  # NOTE: fter rebooting we need to run this file again
+  sudo reboot
+fi
+
+# Check if nvidia driver is installed
+if ! [ -f /usr/bin/nvidia-smi ]; then
+  echo "nvidia driver is not installed, installing"
+  # Install NVIDIA Linux toolkit 510.54
+  wget https://us.download.nvidia.com/XFree86/Linux-x86_64/510.54/NVIDIA-Linux-x86_64-510.54.run
+  chmod +x NVIDIA-Linux-x86_64-510.54.run
+  sudo bash ./NVIDIA-Linux-x86_64-510.54.run
+  rm NVIDIA-Linux-x86_64-510.54.run
+fi
 
 # Install CUDA toolkit 11.3 Upgrade 1 for Ubuntu 20.04 LTS
 wget https://developer.download.nvidia.com/compute/cuda/repos/ubuntu2004/x86_64/cuda-ubuntu2004.pin
@@ -34,10 +43,10 @@ sudo apt-get -y install cuda
 # Install cuDNN 8.2.1 for CUDA 11.x
 # authed download, so just ended up rsyncing it up to the machine from my local
 # wget https://developer.nvidia.com/compute/machine-learning/cudnn/secure/8.2.1.32/11.3_06072021/Ubuntu20_04-x64/libcudnn8_8.2.1.32-1+cuda11.3_amd64.deb
-sudo dpkg -i libcudnn8_8.2.1.32-1+cuda11.3_amd64.deb
-sudo apt-get update
-sudo apt-get -y install libcudnn8
-rm libcudnn8_8.2.1.32-1+cuda11.3_amd64.deb
+# sudo dpkg -i libcudnn8_8.2.1.32-1+cuda11.3_amd64.deb
+# sudo apt-get update
+# sudo apt-get -y install libcudnn8
+# rm libcudnn8_8.2.1.32-1+cuda11.3_amd64.deb
 
 # Install Miniconda
 wget https://repo.anaconda.com/miniconda/Miniconda3-py39_4.9.2-Linux-x86_64.sh
